# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CasePriority": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "CasePriority",
        "type": "string"
      },
...
      "CreateCaseFromInvestigationResult": {
        "additionalProperties": false,
        "properties": {
          "case": {
            "$ref": "#/components/schemas/FraudCase"
          },
          "investigation": {
            "$ref": "#/components/schemas/InvestigationCase"
          },
          "linked_alerts": {
            "items": {
              "$ref": "#/components/schemas/FraudAlert"
            },
            "title": "Linked Alerts",
            "type": "array"
          }
        },
        "required": [
          "investigation",
          "case",
          "linked_alerts"
        ],
        "title": "CreateCaseFromInvestigationResult",
        "type": "object"
      },
...
      "FraudAlert": {
        "additionalProperties": false,
        "description": "An auto-generated or manually created alert that may lead to a case.",
        "properties": {
          "acknowledged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "alert_id": {
            "title": "Alert Id",
            "type": "string"
          },
          "assigned_analyst_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Id"
          },
          "assigned_analyst_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "linked_case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linked Case Id"
          },
          "narrative": {
            "title": "Narrative",
            "type": "string"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "rule_code": {
            "title": "Rule Code",
            "type": "string"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "severity": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/WorkflowSourceType",
            "default": "scenario"
          },
          "status": {
            "$ref": "#/components/schemas/AlertStatus",
            "default": "new"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "alert_id",
          "source_id",
          "rule_code",
          "title",
          "severity",
          "narrative",
          "created_at"
        ],
        "title": "FraudAlert",
        "type": "object"
      },
      "FraudCase": {
        "additionalProperties": false,
        "description": "A persistent fraud investigation case with lifecycle management.",
        "properties": {
          "alert_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Alert Count",
            "type": "integer"
          },
          "assigned_analyst_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Id"
          },
          "assigned_analyst_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Name"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "comment_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Comment Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "disposition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseDisposition"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "$ref": "#/components/schemas/CasePriority",
            "default": "medium"
          },
          "resolution_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Notes"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "risk_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "type": "integer"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "sla_deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Deadline"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/WorkflowSourceType",
            "default": "scenario"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus",
            "default": "open"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "source_id",
          "title",
          "risk_score",
          "risk_level",
          "summary",
          "created_at",
          "updated_at"
        ],
        "title": "FraudCase",
        "type": "object"
      },
      "FraudScenario": {
        "additionalProperties": false,
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountProfile"
            },
            "title": "Accounts",
            "type": "array"
          },
          "customers": {
            "items": {
              "$ref": "#/components/schemas/CustomerProfile"
            },
            "title": "Customers",
            "type": "array"
          },
          "devices": {
            "items": {
              "$ref": "#/components/schemas/DeviceProfile"
            },
            "title": "Devices",
            "type": "array"
          },
          "hypothesis": {
            "title": "Hypothesis",
            "type": "string"
          },
          "industry": {
            "title": "Industry",
            "type": "string"
          },
          "investigator_notes": {
            "items": {
              "$ref": "#/components/schemas/InvestigatorNote"
            },
            "title": "Investigator Notes",
            "type": "array"
          },
          "merchants": {
            "items": {
              "$ref": "#/components/schemas/MerchantProfile"
            },
            "title": "Merchants",
            "type": "array"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/ScenarioTag"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionRecord"
            },
            "title": "Transactions",
            "type": "array"
          }
        },
        "required": [
          "scenario_id",
          "title",
          "industry",
          "summary",
          "hypothesis",
          "tags",
          "customers",
          "accounts",
          "devices",
          "merchants",
          "transactions",
          "investigator_notes"
        ],
        "title": "FraudScenario",
        "type": "object"
      },
...
      "GetScenarioResult": {
        "additionalProperties": false,
        "properties": {
          "scenario": {
            "$ref": "#/components/schemas/FraudScenario"
          }
        },
        "required": [
          "scenario"
        ],
        "title": "GetScenarioResult",
        "type": "object"
      },
...
      "GraphAnalysisResult": {
        "additionalProperties": false,
        "description": "Results from graph-based relationship analysis.",
        "properties": {
          "community_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Community Count",
            "type": "integer"
          },
          "connected_components": {
            "minimum": 0.0,
            "title": "Connected Components",
            "type": "integer"
          },
          "density": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Density",
            "type": "number"
          },
          "highest_degree_entity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntityReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "highest_degree_score": {
            "default": 0,
            "minimum": 0.0,
            "title": "Highest Degree Score",
            "type": "integer"
          },
          "hub_entities": {
            "items": {
              "$ref": "#/components/schemas/EntityReference"
            },
            "title": "Hub Entities",
            "type": "array"
          },
          "risk_amplification_factor": {
            "default": 1.0,
            "minimum": 1.0,
            "title": "Risk Amplification Factor",
            "type": "number"
          },
          "shortest_path_length": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shortest Path Length"
          }
        },
        "required": [
          "connected_components",
          "density"
        ],
        "title": "GraphAnalysisResult",
        "type": "object"
      },
      "GraphLink": {
        "additionalProperties": false,
        "properties": {
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "relation": {
            "title": "Relation",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/EntityReference"
          },
          "target": {
            "$ref": "#/components/schemas/EntityReference"
          }
        },
        "required": [
          "relation",
          "source",
          "target",
          "explanation"
        ],
        "title": "GraphLink",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvestigateScenarioCommand": {
        "additionalProperties": false,
        "properties": {
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          }
        },
        "required": [
          "scenario_id"
        ],
        "title": "InvestigateScenarioCommand",
        "type": "object"
      },
      "InvestigateScenarioResult": {
        "additionalProperties": false,
        "properties": {
          "investigation": {
            "$ref": "#/components/schemas/InvestigationCase"
          }
        },
        "required": [
          "investigation"
        ],
        "title": "InvestigateScenarioResult",
        "type": "object"
      },
      "InvestigationCase": {
        "additionalProperties": false,
        "properties": {
          "graph_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphAnalysisResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "graph_links": {
            "items": {
              "$ref": "#/components/schemas/GraphLink"
            },
            "title": "Graph Links",
            "type": "array"
          },
          "investigation_leads": {
            "items": {
              "$ref": "#/components/schemas/InvestigationLead"
            },
            "title": "Investigation Leads",
            "type": "array"
          },
          "metrics": {
            "$ref": "#/components/schemas/InvestigationMetrics"
          },
          "provider_summary": {
            "$ref": "#/components/schemas/ProviderSummary"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "title": "Recommended Actions",
            "type": "array"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "scenario": {
            "$ref": "#/components/schemas/ScenarioOverview"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "suspicious_transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionRecord"
            },
            "title": "Suspicious Transactions",
            "type": "array"
          },
          "text_signals": {
            "items": {
              "$ref": "#/components/schemas/TextSignal"
            },
            "title": "Text Signals",
            "type": "array"
          },
          "top_rule_hits": {
            "items": {
              "$ref": "#/components/schemas/RuleHit"
            },
            "title": "Top Rule Hits",
            "type": "array"
          },
          "total_risk_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Total Risk Score",
            "type": "integer"
          }
        },
        "required": [
          "scenario",
          "risk_level",
          "total_risk_score",
          "summary",
          "metrics",
          "provider_summary",
          "top_rule_hits",
          "graph_links",
          "text_signals"
        ],
        "title": "InvestigationCase",
        "type": "object"
      },
      "InvestigationLead": {
        "additionalProperties": false,
        "description": "An investigation-ready hypothesis assembled from related findings.",
        "properties": {
          "entities": {
            "items": {
              "$ref": "#/components/schemas/EntityReference"
            },
            "title": "Entities",
            "type": "array"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "hypothesis": {
            "title": "Hypothesis",
            "type": "string"
          },
          "lead_id": {
            "title": "Lead Id",
            "type": "string"
          },
          "lead_type": {
            "title": "Lead Type",
            "type": "string"
          },
          "narrative": {
            "title": "Narrative",
            "type": "string"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "title": "Recommended Actions",
            "type": "array"
          },
          "severity": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "supporting_anomaly_ids": {
            "items": {
              "type": "string"
            },
            "title": "Supporting Anomaly Ids",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "lead_id",
          "lead_type",
          "title",
          "severity",
          "hypothesis",
          "narrative"
        ],
        "title": "InvestigationLead",
        "type": "object"
      },
      "InvestigationMetrics": {
        "additionalProperties": false,
        "properties": {
          "linked_customer_count": {
            "minimum": 0.0,
            "title": "Linked Customer Count",
            "type": "integer"
          },
          "shared_device_count": {
            "minimum": 0.0,
            "title": "Shared Device Count",
            "type": "integer"
          },
          "suspicious_transaction_count": {
            "minimum": 0.0,
            "title": "Suspicious Transaction Count",
            "type": "integer"
          },
          "suspicious_transaction_volume": {
            "minimum": 0.0,
            "title": "Suspicious Transaction Volume",
            "type": "number"
          },
          "total_transaction_volume": {
            "minimum": 0.0,
            "title": "Total Transaction Volume",
            "type": "number"
          }
        },
        "required": [
          "total_transaction_volume",
          "suspicious_transaction_volume",
          "suspicious_transaction_count",
          "shared_device_count",
          "linked_customer_count"
        ],
        "title": "InvestigationMetrics",
        "type": "object"
      },
...
      "ListScenariosResult": {
        "additionalProperties": false,
        "properties": {
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioOverview"
            },
            "title": "Scenarios",
            "type": "array"
          }
        },
        "required": [
          "scenarios"
        ],
        "title": "ListScenariosResult",
        "type": "object"
      },
...
      "ProviderSummary": {
        "additionalProperties": false,
        "properties": {
          "active_reasoning_provider": {
            "title": "Active Reasoning Provider",
            "type": "string"
          },
          "active_text_provider": {
            "title": "Active Text Provider",
            "type": "string"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "requested_reasoning_provider": {
            "title": "Requested Reasoning Provider",
            "type": "string"
          },
          "requested_text_provider": {
            "title": "Requested Text Provider",
            "type": "string"
          }
        },
        "required": [
          "requested_reasoning_provider",
          "active_reasoning_provider",
          "requested_text_provider",
          "active_text_provider"
        ],
        "title": "ProviderSummary",
        "type": "object"
      },
      "RiskLevel": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
...
      "ScenarioOverview": {
        "additionalProperties": false,
        "properties": {
          "baseline_risk": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "hypothesis": {
            "title": "Hypothesis",
            "type": "string"
          },
          "industry": {
            "title": "Industry",
            "type": "string"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/ScenarioTag"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "total_volume": {
            "minimum": 0.0,
            "title": "Total Volume",
            "type": "number"
          },
          "transaction_count": {
            "minimum": 0.0,
            "title": "Transaction Count",
            "type": "integer"
          }
        },
        "required": [
          "scenario_id",
          "title",
          "industry",
          "summary",
          "hypothesis",
          "tags",
          "transaction_count",
          "total_volume",
          "baseline_risk"
        ],
        "title": "ScenarioOverview",
        "type": "object"
      },
...
      "TextSignal": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "signal_id": {
            "title": "Signal Id",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_kind": {
            "$ref": "#/components/schemas/TextSignalKind"
          }
        },
        "required": [
          "signal_id",
          "provider",
          "source_kind",
          "source_id",
          "label",
          "confidence",
          "rationale"
        ],
        "title": "TextSignal",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "A fraud triage workspace for uploaded transaction datasets, with persistent alerts and cases, plus reference scenario investigations for validation and rule calibration. Statistical and behavioral analysis remain the source of truth, while the optional copilot layer explains the results in operator-facing language.",
    "title": "Relational Fraud Intelligence",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/investigations": {
      "post": {
        "description": "Executes a full fraud investigation on a reference scenario: text signal extraction, rule-based risk reasoning, graph analysis, and case assembly.",
        "operationId": "investigate_scenario_api_v1_investigations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigateScenarioCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigateScenarioResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run fraud investigation",
        "tags": [
          "Investigations"
        ]
      }
    },
    "/api/v1/investigations/{scenario_id}/case": {
      "post": {
        "description": "Runs the scenario investigation, creates a persistent fraud case from the highest-signal scenario leads, and links any open scenario alerts to that case.",
        "operationId": "create_case_from_investigation_api_v1_investigations__scenario_id__case_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCaseFromInvestigationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a case from a scenario investigation",
        "tags": [
          "Investigations"
        ]
      }
    },
    "/api/v1/scenarios": {
      "get": {
        "description": "Returns all seeded fraud scenarios with baseline risk assessment.",
        "operationId": "list_scenarios_api_v1_scenarios_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListScenariosResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List fraud scenarios",
        "tags": [
          "Investigations"
        ]
      }
    },
    "/api/v1/scenarios/{scenario_id}": {
      "get": {
        "description": "Returns full scenario data including customers, accounts, devices, merchants, and transactions.",
        "operationId": "get_scenario_api_v1_scenarios__scenario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetScenarioResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get scenario details",
        "tags": [
          "Investigations"
        ]
      }
    },
...
  },
...
}